Fixed "CalculatorLibrary" folder structure
