+++ Calculator.TomDonegan/Calculator.TomDonegan/Calculator.TomDonegan.csproj
@@ -1,0 +1,20 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
+    <GenerateTargetFrameworkAttribute>false</GenerateTargetFrameworkAttribute>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net7.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Remove="Helpers.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="Calculator.TomDonegan\CalculatorLibrary\CalculatorLibrary.csproj" />
+  </ItemGroup>
+
+</Project>
