+++ Calculator.TomDonegan/Calculator.TomDonegan/Calculator.TomDonegan.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.6.33829.357
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calculator.TomDonegan", "Calculator.TomDonegan.csproj", "{C75281A7-F577-41B3-8124-9CFE7DB82B48}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Calculator.TomDonegan", "Calculator.TomDonegan.csproj", "{C75281A7-F577-41B3-8124-9CFE7DB82B48}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CalculatorLibrary", "CalculatorLibrary\CalculatorLibrary.csproj", "{497BA0BC-9895-4C92-8B91-8FE94115F150}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{C75281A7-F577-41B3-8124-9CFE7DB82B48}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C75281A7-F577-41B3-8124-9CFE7DB82B48}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C75281A7-F577-41B3-8124-9CFE7DB82B48}.Release|Any CPU.Build.0 = Release|Any CPU
+		{497BA0BC-9895-4C92-8B91-8FE94115F150}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{497BA0BC-9895-4C92-8B91-8FE94115F150}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{497BA0BC-9895-4C92-8B91-8FE94115F150}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{497BA0BC-9895-4C92-8B91-8FE94115F150}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
