+++ Calculator.TomDonegan/Calculator.TomDonegan/Notes.txt
@@ -1,0 +1,42 @@
+﻿Future Notes:
+	Queue<CalculationHistory> history = new Queue<CalculationHistory>();
+    
+    foreach ( var calculation in history)
+                {
+                    Console.WriteLine($"{calculation.calculationID}: {calculation.firstNumber} {calculation.stringOperation} {calculation.secondNumber} = {calculation.calculationResult}");
+                } 
+
+    history.Enqueue(new CalculationHistory(numberOfUses, cleanNum1, cleanNum2, op, result));
+
+    public class CalculationHistory
+    {
+        public int calculationID { get; set; }
+        public double firstNumber { get; set; }
+        public double secondNumber { get; set; }
+        public string stringOperation { get; set; }
+
+        public double calculationResult { get; set; }
+        public CalculationHistory(int ID, double firstNumber, double secondNumber, string op, double result) 
+        {
+            switch (op) 
+            {
+                case "a":
+                    op = "+";
+                    break;
+                case "s":
+                    op = "-";
+                    break;
+                case "m":
+                    op = "*";
+                    break;
+                case "d":
+                    op = "/";
+                    break;
+            }
+            this.calculationID = ID;
+            this.firstNumber = firstNumber;
+            this.secondNumber = secondNumber;
+            this.stringOperation = op;
+            this.calculationResult = result;
+        }
+    }
